时间功能修改

+++ LuaTest/main.cpp
@@ -12,7 +12,7 @@
 
 class TestTimer: public IFxTimer
 {
-	virtual bool OnTimer(float fSecond)
+	virtual bool OnTimer(double fSecond)
 	{
 		LogFun(LT_File | LT_Screen, LogLv_Error, "%f", fSecond);
 		return true;
